--- src/pages/MerchantVoucherDashboardV2/index.ts
@@ -1,1 +1,0 @@
-export { default } from './MerchantVoucherDashboardV2';
